Dev to Dev

See merge request products/deepin-draw!6

+++ src/frame/toptoolbar.cpp
@@ -220,7 +220,7 @@ void TopToolbar::slotZoom(const qreal &scale)
 
 void TopToolbar::slotSetScale(const qreal scale)
 {
-    QString strScale = QString::number(scale * 100) + "%";
+    QString strScale = QString::number(int(scale * 100)) + "%";
     m_zoomMenuComboBox->setMenuButtonTextAndIcon(strScale, QIcon());
 }
